+++ include/debug_free.h
@@ -1,11 +1,23 @@
-#ifndef DEBUG_FREE_H
-#define DEBUG_FREE_H
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   debug_free.h                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 02:42:51 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 02:42:55 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-#include <stdio.h>
-#include <stdlib.h>
+// #ifndef DEBUG_FREE_H
+// #define DEBUG_FREE_H
 
-void my_free(void **ptr, const char *file, int line);
+// #include <stdio.h>
+// #include <stdlib.h>
 
-#define free(ptr) my_free(ptr, __FILE__, __LINE__)
+// void my_free(void **ptr, const char *file, int line);
 
-#endif // DEBUG_FREE_H
+// #define free(ptr) my_free(ptr, __FILE__, __LINE__)
+
+// #endif // DEBUG_FREE_H
