+++ include/debug_helpers.h
@@ -1,4 +1,15 @@
-/* debug_helpers.h - debug用の関数を定義するヘッダファイル */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   debug_helpers.h                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 02:46:56 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 02:46:56 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef DEBUG_HELPERS_H
 # define DEBUG_HELPERS_H
