+++ include/free.h
@@ -1,4 +1,15 @@
-/* free.h - メモリ解放関連のヘッダーファイル */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   free.h                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 02:51:36 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 02:51:36 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FREE_H
 # define FREE_H
