+++ include/utils.h
@@ -1,31 +1,42 @@
-/* utils.h - libft関数。 */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 02:48:04 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 02:48:04 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef UTILS_H
 # define UTILS_H
 
 # include <stddef.h>
 
-/* utils.c */
+/* util_alloc.c */
 void	ft_bzero(void *s, size_t n);
-size_t	ft_strlcat(char *dest, const char *src, size_t size);
-size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-char	*ft_strchr(const char *s, int c);
-int		ft_isdigit(int c);
-
-/* builtin.c ? */
-size_t	ft_strlen(const char *str);
-char	*ft_strndup(char *src, long len);
-char	*ft_strdup(char *src);
-
-// calloc
 void	*ft_calloc(size_t count, size_t size);
-// itoa
 char	*ft_itoa(int n);
 
-char	*ft_strncpy(char *dest, const char *src, size_t n);
+/* util_cmp.c */
+int		ft_isdigit(int c);
+int		ft_strcmp(const char *s1, const char *s2);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
+
+/* util_mem.c */
+size_t	ft_strlcpy(char *dest, const char *src, size_t size);
+size_t	ft_strlcat(char *dest, const char *src, size_t size);
 char	*ft_strcpy(char *dest, const char *src);
+char	*ft_strncpy(char *dest, const char *src, size_t n);
+
+/* util_str.c */
+size_t	ft_strlen(const char *str);
+char	*ft_strchr(const char *s, int c);
+char	*ft_strdup(char *src);
+char	*ft_strndup(char *src, long len);
 
 char	**ft_split(char const *s, char c);
 
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
-int		ft_strcmp(const char *s1, const char *s2);
 #endif
