+++ include/xlibc.h
@@ -1,4 +1,15 @@
-/* xlibc.h - 標準Cライブラリのラッパー関数 */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   xlibc.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 02:44:57 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 02:44:58 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef XLIBC_H
 # define XLIBC_H
