+++ src/error.c
@@ -1,4 +1,15 @@
-/* error.c - エラー処理に関する関数の定義。 */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   error.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 01:11:11 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 01:15:45 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "error.h"
 #include "minishell.h"
 #include "utils.h"
@@ -12,7 +23,6 @@ void	memory_error_and_exit(char *func_name)
 	exit(SC_FATAL_ERROR);
 }
 
-// TODO: デバッグ用の関数
 void	assert_error(const char *msg, char *location)
 {
 	write(STDERR_FILENO, "Assert Error: ", 14);
@@ -34,42 +44,6 @@ void	error_exit(const char *msg, int exit_status)
 		write(STDERR_FILENO, msg, ft_strlen(msg));
 	write(STDERR_FILENO, "\n", 1);
 	exit(exit_status);
-}
-
-static const char	*get_token_string(t_token *token)
-{
-	if (token->type == TK_EOF)
-		return ("newline");
-	if ((token->type == TK_WORD || token->type == TK_SQUOTE
-			|| token->type == TK_DQUOTE) && token->word)
-		return ((const char *)(token->word));
-	if (token->type == TK_IO_NUM)
-		return ("IO number");
-	if (token->type == TK_PIPE)
-		return ("|");
-	if (token->type == TK_REDIR_IN)
-		return ("<");
-	if (token->type == TK_REDIR_OUT)
-		return (">");
-	if (token->type == TK_HEREDOC)
-		return ("<<");
-	if (token->type == TK_APPEND)
-		return (">>");
-	if (token->type == TK_BLANK)
-		return ("blank");
-	return ("unknown or undefined token");
-}
-
-void	parser_error(t_token *unexpected_token)
-{
-	const char	*token_str = get_token_string(unexpected_token);
-
-	write(STDERR_FILENO, "minishell: syntax error near unexpected token `", 47);
-	if (unexpected_token)
-	{
-		write(STDERR_FILENO, token_str, ft_strlen(token_str));
-	}
-	write(STDERR_FILENO, "'\n", 2);
 }
 
 void	report_error(char *cmd, char *file, char *info)
@@ -106,8 +80,8 @@ void	sys_error(char *cmd, char *msg)
 	if (msg && *msg)
 	{
 		write(STDERR_FILENO, msg, ft_strlen(msg));
-
-	}write(STDERR_FILENO, ": ", 2);
+	}
+	write(STDERR_FILENO, ": ", 2);
 	write(STDERR_FILENO, strerror(e), ft_strlen(strerror(e)));
 	write(STDERR_FILENO, "\n", 1);
 }
