+++ src/parser_init.c
@@ -1,4 +1,15 @@
-/* parser_init.c - パーサーの初期化関数の定義 */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parser_init.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/09/11 01:50:25 by sakitaha          #+#    #+#             */
+/*   Updated: 2024/09/11 01:50:28 by sakitaha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "free.h"
 #include "parser.h"
 #include "utils.h"
